glsl: fix indexing order of 2D arrays
My original implementation of 2D arrays switched the vertex and attribute
indices in dereferencing.  This commit fixes that, and as a side effect makes
the code somewhat cleaner.

This could be combined with "glsl: support compilation of geometry shaders
via EXT/ARB_geometry_shader4" to make the patch set cleaner.